Fence and rain made
Imported my fence which will act as an environment tool as well as an obstacle for the player to jump over.
I then made a rain particle system with a custom shader to dither the rain to create fake transparency which will help with performance. a 2000 particle rain system is currently adding 2 draw calls and doesnt seem to be taking away any fps.