Dirt path imported and partly painted
I made and imported a dirt path into the scene. Now the ground is completely done. We have a grass shader, a grass texture to go underneath and now a dirt path to guide the player. The only things next to do is some decals to guide the players in dashing/walking. However this will only come later if i have time.